requirements.txt
bert-serving-client
GitPython>=3.1,<3.2
